fix: capture parentID before same-name dedupe; split files for lint
upsertAgent removes any agent with the same name+wsID before writing
the new entry, which was clobbering the enriched parentID we tried to
preserve. Read the existing parentID first.

Also splits MenuState.swift, VPNMenuItem.swift, and VPNMenuStateTests.swift
to satisfy file_length and type_body_length lint rules.
