+++ Coder-Desktop/Coder-Desktop/VPN/MenuState.swift
@@ -159,66 +159,6 @@ struct Workspace: Identifiable, Equatable, Comparable {
     }
 }
 
-/// WorkspaceGroup is the unit the tray renders: one workspace with the agents
-/// belonging to it, organized into a parent/child tree.
-struct WorkspaceGroup: Identifiable, Equatable, Comparable {
-    let workspace: Workspace
-    /// All agents in the workspace, unsorted.
-    let agents: [Agent]
-
-    var id: UUID {
-        workspace.id
-    }
-
-    /// Aggregate status: worst-of among agents, or .off if the workspace has
-    /// none (treated as offline).
-    var status: AgentStatus {
-        agents.map(\.status).max() ?? .off
-    }
-
-    /// Top-level agents are those without a parent, plus any agent whose
-    /// parentID points outside this group's known agents (orphans render at the
-    /// top level rather than disappearing).
-    var topLevelAgents: [Agent] {
-        let knownIDs = Set(agents.map(\.id))
-        return agents
-            .filter { agent in
-                guard let parentID = agent.parentID else { return true }
-                return !knownIDs.contains(parentID)
-            }
-            .sorted()
-    }
-
-    /// Children of the given parent agent, sorted.
-    func children(of parentID: UUID) -> [Agent] {
-        agents.filter { $0.parentID == parentID }.sorted()
-    }
-
-    /// Tree-walked agents paired with their indent depth (1 for top-level).
-    /// Handles arbitrary parent_id depth; cycles are guarded against by
-    /// tracking visited IDs.
-    var indentedAgents: [(agent: Agent, indent: Int)] {
-        var result: [(Agent, Int)] = []
-        var visited: Set<UUID> = []
-        func walk(_ agent: Agent, indent: Int) {
-            guard visited.insert(agent.id).inserted else { return }
-            result.append((agent, indent))
-            for child in children(of: agent.id) {
-                walk(child, indent: indent + 1)
-            }
-        }
-        for top in topLevelAgents {
-            walk(top, indent: 1)
-        }
-        return result
-    }
-
-    static func < (lhs: WorkspaceGroup, rhs: WorkspaceGroup) -> Bool {
-        if lhs.status != rhs.status { return lhs.status < rhs.status }
-        return lhs.workspace < rhs.workspace
-    }
-}
-
 struct VPNMenuState {
     var agents: [UUID: Agent] = [:]
     var workspaces: [UUID: Workspace] = [:]
@@ -248,6 +188,11 @@ struct VPNMenuState {
         // Remove trailing dot if present
         let nonEmptyHosts = agent.fqdn.map { $0.hasSuffix(".") ? String($0.dropLast()) : $0 }
 
+        // The proto doesn't carry parent_id, so preserve any value we already
+        // enriched for this agent from the HTTP API. Captured here because the
+        // same-name dedupe below removes the existing entry.
+        let existingParentID = agents[id]?.parentID
+
         // An existing agent with the same name, belonging to the same workspace
         // is from a previous workspace build, and should be removed.
         agents.filter { $0.value.name == agent.name && $0.value.wsID == wsID }
@@ -267,9 +212,6 @@ struct VPNMenuState {
                     : nil
             )
         }
-        // The proto doesn't carry parent_id, so preserve any value we already
-        // enriched for this agent from the HTTP API.
-        let existingParentID = agents[id]?.parentID
         agents[id] = Agent(
             id: id,
             name: agent.name,
